move typed getters to Property & drop fallback interface
Better usability for user code if they deserialize individual Propertys.

The fallback variants are not used internally and are redundant
because there is a separate default behavior.

Removed the msg parameter for Property exceptions.
It was only used by a single edge-case which didn't really justify
the use of the exception anyway.
Instead allow `undefined` to be instantiated without parameters
because the Property does not know its own name.